rename to batch_generator
