+++ pyveda/frameworks/batch_generator.py
